requirements.txt
# Django и основные зависимости
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# База данных
psycopg2-binary==2.9.7

# HTTP запросы
requests==2.31.0

# XML обработка
lxml==4.9.3

# Безопасность
cryptography==41.0.7

# Статические файлы
whitenoise==6.6.0

# Производство
gunicorn==21.2.0
supervisor==4.2.5

# Мониторинг
sentry-sdk==1.38.0

# Кэширование
redis==5.0.1
django-redis==5.4.0

# Логирование
python-logging-loki==0.3.1

# Утилиты
python-decouple==3.8
Pillow==10.1.0
